prov/util,tcp,rxm,rxd,netdir,sockets,hook: Fix Windows build warnings
A previous patch (commit 0264b9daa153) eliminated a great amount of
warnings caused by headers. However, there are still many warnings
caused by source code of individual providers. That's what this patch
tries to address.

Fix 400+ warnings scattered across 50+ files. Most of the warnings are
either implicit type conversion from larger type size to smaller type
size or comparison between signed & unsigned values.

There are still a few types of warnings being left unhandled:

(1) assignment in condition in 'while ((a = func()) { ... }' structure.
    don't see a good replacement for that. similar complains for 'if'
    statements have been fixed.

(2) windows socket fd type 'SOCKET' is different from 'int'. That requires
    more work to get it right.

(3) structure padding due to alignment specifier.

(4) using enum for bit field base type is non-standard extension.

(5) implicit type conversion from 'int' to 'float' in the atomic macros,
    which are quite complicated to decode.

+++ include/ofi_osd.h
@@ -85,7 +85,7 @@ enum {
 	OFI_ENDIAN_LITTLE_WORD,	/* Middle-endian, PDP-11 style */
 };
 
-static inline int ofi_detect_endianness(void)
+static inline uint8_t ofi_detect_endianness(void)
 {
 	union {
 		uint8_t data[sizeof(uint32_t)];
